moved searchApp to the navbar
