+++ src/components/App.js
@@ -2,7 +2,7 @@ import React, { Component } from "react"
 import NavBar from "./navbar/NavBar"
 import { BrowserRouter as Router } from 'react-router-dom'
 import ApplicationViews from "./ApplicationViews"
-import SearchApp from "./navbar/SearchApp"
+
 
 import "./App.css"
 import "bootstrap/dist/css/bootstrap.min.css"
@@ -13,7 +13,7 @@ class App extends Component {
     return (
       <Router>
         <React.Fragment>
-          <SearchApp />
+          <NavBar />
           <ApplicationViews />
         </React.Fragment>
       </Router>
